refactor(daemon): split rp_broker module
